fe_stream/requirements.txt
requests==2.31.0
pillow==10.1.0
streamlit==1.31.1